refactor: move verification helpers to separate module
