+++ ollamarama.py
@@ -11,12 +11,8 @@ from nio import (
     MatrixRoom,
     RoomMessageText,
     KeyVerificationEvent,
-    KeyVerificationStart,
-    KeyVerificationKey,
-    KeyVerificationMac,
-    KeyVerificationCancel,
-    ToDeviceMessage,
 )
+from verification import VerificationMixin
 import json
 import os
 import datetime
@@ -26,7 +22,7 @@ import markdown
 import logging
 import logging.config
 
-class ollamarama:
+class ollamarama(VerificationMixin):
     """
     An Ollama-based chatbot for the Matrix chat protocol, supporting dynamic personalities, 
     custom prompts, and cross-user interactions.
@@ -126,19 +122,6 @@ class ollamarama:
             ignore_unverified_devices=True,
         )
 
-    async def allow_devices(self, user_id):
-        """Ensure messages can be sent to all of a user's devices."""
-        try:
-            for device in self.client.device_store.active_user_devices(user_id):
-                try:
-                    if self.client.olm.is_device_blacklisted(device):
-                        self.client.unblacklist_device(device)
-                    if not self.client.olm.is_device_verified(device):
-                        self.client.ignore_device(device)
-                except Exception as e:
-                    self.log(f"Failed to allow device {device.id} for {user_id}: {e}")
-        except Exception as e:
-            self.log(f"Failed to allow devices for {user_id}: {e}")
 
     async def add_history(self, role, channel, sender, message):
         """
@@ -409,62 +392,6 @@ class ollamarama:
                 except:
                     pass
 
-    async def emoji_verification_callback(self, event):
-        """Auto-accept incoming emoji verification requests (SAS)."""
-        client = self.client
-        try:
-            if isinstance(event, KeyVerificationStart):
-                if "emoji" not in event.short_authentication_string:
-                    return
-                await client.accept_key_verification(event.transaction_id)
-                sas = client.key_verifications[event.transaction_id]
-                todevice_msg = sas.share_key()
-                await client.to_device(todevice_msg)
-            elif isinstance(event, KeyVerificationKey):
-                sas = client.key_verifications[event.transaction_id]
-                emojis = sas.get_emoji()
-                self.log(f"Emoji verification requested: {emojis}")
-                # Auto-accept for bot
-                await client.confirm_short_auth_string(event.transaction_id)
-            elif isinstance(event, KeyVerificationMac):
-                sas = client.key_verifications[event.transaction_id]
-                try:
-                    done = ToDeviceMessage(
-                        "m.key.verification.done",
-                        event.sender,
-                        sas.other_olm_device.id,
-                        {"transaction_id": event.transaction_id},
-                    )
-                    await client.to_device(done)
-                    self.log("Emoji verification was successful.")
-                except Exception:
-                    self.log("Failed to complete emoji verification.")
-            elif isinstance(event, KeyVerificationCancel):
-                self.log(f"Verification cancelled.")
-        except Exception:
-            self.log("Exception during emoji verification.")
-
-    async def log_to_device_event(self, event):
-        # Accept verification requests so the emoji flow can proceed
-        if hasattr(event, 'type') and event.type == "m.key.verification.request":
-            try:
-                txn_id = event.source['content']['transaction_id']
-                from_device = event.source['content']['from_device']
-                self.log("Verification ready message sent.")
-                content = {
-                    "from_device": self.device_id,
-                    "methods": ["m.sas.v1"],
-                    "transaction_id": txn_id
-                }
-                message = ToDeviceMessage(
-                    "m.key.verification.ready",
-                    event.sender,
-                    from_device,
-                    content,
-                )
-                await self.client.to_device(message)
-            except Exception:
-                self.log("Failed to send verification ready message.")
 
     async def main(self):
         """
